Change light intensity type to uint32_t
The brightest sunlight can result in at most 120,000 lux. We want to be able to represent that value in our system.

+++ src/app/port-interfaces/hal/light_intensity.h
@@ -17,10 +17,11 @@ extern "C"
  *   - 0 lx → 0
  *   - 100 lx → 100
  *   - 12345 lx → 12345
+ *   - 100000 lx → 100000
  *
- * Light intensity values range allowed by representation (in lx): 0 to 65535.
+ * Light intensity values range allowed by representation (in lx): 0 to 4294967295.
  */
-typedef uint16_t LightIntensity;
+typedef uint32_t LightIntensity;
 
 #ifdef __cplusplus
 }
